requirements.txt
absl-py==0.9.0
aiohttp==3.6.2
appnope==0.1.0
APScheduler==3.6.3
arrow==0.15.5
asgiref==3.2.7
astor==0.8.1
async-timeout==3.0.1
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.0
bleach==3.1.4
blessed==1.17.4
bokeh==2.0.2
Bottleneck==1.2.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cloudpickle==1.4.1
colorama==0.4.3
cycler==0.10.0
dask==2.15.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
distributed==2.15.2
Django==3.0.5
django-apscheduler==0.3.0
django-crontab==0.7.1
django-extensions==2.2.9
django-import-export==2.1.0
django-picklefield==2.1.1
django-q==1.2.1
django-storages==1.6.6
djangorestframework==3.11.0
entrypoints==0.3
filelock==3.0.12
fsspec==0.7.3
future==0.18.2
gast==0.2.2
google==2.0.3
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
HeapDict==1.0.1
html5lib==1.0.1
idna==2.8
importlib-metadata==1.6.0
ipykernel==5.2.1
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.0
jedi==0.17.0
Jinja2==2.11.2
joblib==0.14.1
json5==0.9.4
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
llvmlite==0.32.0
locket==0.2.0
lxml==4.3.2
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
modin==0.7.3
msgpack==1.0.0
multidict==4.7.5
mysqlclient==1.4.6
nbconvert==5.6.1
nbformat==5.0.5
notebook==5.7.8
numba==0.49.0
numpy==1.18.2
oauthlib==3.1.0
opt-einsum==3.2.1
packaging==20.3
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
partd==1.1.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pip==20.1
prometheus-client==0.7.1
prompt-toolkit==1.0.18
protobuf==3.11.3
psutil==5.7.0
ptyprocess==0.6.0
py-spy==0.3.3
pyarrow==0.16.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
PyMySQL==0.9.3
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
pyzmq==19.0.0
qtconsole==4.7.3
QtPy==1.9.0
ray==0.8.4
redis==3.5.0
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
setuptools==46.1.3
simplegeneric==0.8.1
six==1.14.0
sortedcontainers==2.1.0
soupsieve==2.0
SQLAlchemy==1.3.7
sqlparse==0.3.1
swifter==0.304
TA-Lib==0.4.17
tablib==1.1.0
tblib==1.6.0
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
toolz==0.10.0
torch==1.5.0
torchvision==0.6.0
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
typing-extensions==3.7.4.2
tzlocal==2.0.0
urllib3==1.25.9
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
wheel==0.34.2
widgetsnbextension==3.4.2
wrapt==1.12.1
yarl==1.4.2
zict==2.0.0
zipp==3.1.0